now with missing added files
rcapp.py of receiver
-----------------
- def adjust_threshold (self, ch, fval):
 new function to set the threshold by the transmitter of an channel
   of class PWM_Controller(PCA9685), but not used yet

- method run() of class Observer
  setting faile save after timeout
  then wait until timeout * 10 is pasted and carry out a reboot

Android and PC version of transmitter:
------------------------------------------------
- separation of most code that depends on the mode (Android or Java) in procesing
- improvement of the status indication of the reciever
- draw the IP-addrress of the receiver when touching the indicator icon
  (that made it easier to determine the IP for logging in if any changes in the
   rcmain.py code of the receiver has the be adjust)
- new class  ‘SwitchTh’ extends Switch to change/set the threshold of an channel by an switch
  e.g. 100% or 50 % to move slower
- switch the position of the ON/OFF switch so that the ON position is at the top
- back moving speed to default position improved

+++ sw/Rpi/rcapp.py
@@ -81,7 +81,7 @@ class PWM_Controller(PCA9685):
         self.FSave = []     
         self.imp_tab = []
         self.invert_input = []
-        self.ChTidef = [],[]       
+        self.ChTidef = [],[]  
         # Fills the fail_safe position list with default values
         for i in range(16):
             self.FSave.append(127)           
@@ -89,7 +89,7 @@ class PWM_Controller(PCA9685):
         # Fills the channel time behavior list with default values
         for i in range(16):
             self.ChTidef[self.ix_min].append(self.imp_val(self.i_min))            
-            self.ChTidef[self.ix_diff].append(self.diff_val(self.i_min, self.i_max))   
+            self.ChTidef[self.ix_diff].append(self.diff_val(self.i_min, self.i_max))           
         self.calc_puls_table()
         # update of pulstable for LM298 purposes
         i = 0
@@ -128,7 +128,15 @@ class PWM_Controller(PCA9685):
         self.ChTidef[self.ix_diff][ch] = self.diff_val(min_val, max_val)
         for i in range(255):                 
                 self.imp_tab[ch][i]=(int((self.ChTidef[self.ix_diff][ch] * i/254) \
-                        + self.ChTidef[self.ix_min][ch])) 
+                        + self.ChTidef[self.ix_min][ch]))
+                
+    # ch: channel, fval = Thresshold 0..100 not used
+    def adjust_threshold (self, ch, fval):
+        ndiff =(self.ChTidef[self.ix_diff][ch] * fval/100.0)
+        nmin = ((self.ChTidef[self.ix_min][ch]  \
+                 + (self.ChTidef[self.ix_diff][ch]/2)) - ndiff/2)
+        for i in range(255):                 
+                self.imp_tab[ch][i]=int((ndiff * i/254) + nmin) 
 
     def calc_puls_table_L298 (self, chan):
         """ call after 'calc_puls_table'
@@ -173,9 +181,8 @@ class PWM_Controller(PCA9685):
     def update_app (self, chan, remote_inp):
         if ((not SIM) and (remote_inp == 0)):
             self.fail_safe()
-            time.sleep(1)
             print ("shutdown")
-            time.sleep(2)
+            time.sleep(1)
             os.system("sudo shutdown now")                                    
                                
     def update_L298(self, chan, remote_inp):
@@ -226,7 +233,7 @@ class PWM_Controller(PCA9685):
             elif(hdr == 127):
                 self.trimm_chan(ch, val)
             elif (hdr == 100):
-                self.update_app(ch, val)              
+                self.update_app(ch, val)  
 
 class Observer(Thread):
     ''' Timeout supervision of receiving telegrams.
@@ -239,6 +246,7 @@ class Observer(Thread):
         Thread.__init__(self)
         self.SC = SC
         self.tout = tel_time_out
+        self.tshutdn = tel_time_out * 10.0
         self.ID = ID
         self.start()
 
@@ -246,13 +254,13 @@ class Observer(Thread):
         while True:
             if ((time.time() - self.SC.get_rtime()) > self.tout):
                 self.SC.fail_safe()
-                time.sleep(1)
-                print ('exit' + str(os.getpid()))
-                if SIM:
-                    os.system("sudo kill " + str(os.getpid()))
-                else:
-                    # better rebooting
-                    os.system("sudo reboot")
+                if ((time.time() - self.SC.get_rtime()) > self.tshutdn):
+                    #print ('exit' + str(os.getpid()))
+                    if SIM:
+                        os.system("sudo kill " + str(os.getpid()))
+                    else:
+                        # better rebooting
+                        os.system("sudo reboot")
             time.sleep(1)
